+++ ext/stub.c
@@ -219,6 +219,7 @@ zend_class_entry *stub_issue2651_ce;
 zend_class_entry *stub_issue2654_ce;
 zend_class_entry *stub_issue2656_ce;
 zend_class_entry *stub_issue2666_ce;
+zend_class_entry *stub_issue2674_ce;
 zend_class_entry *stub_issue2679_ce;
 zend_class_entry *stub_issue663_ce;
 zend_class_entry *stub_issue760_ce;
@@ -520,6 +521,7 @@ static PHP_MINIT_FUNCTION(stub)
 	ZEPHIR_INIT(Stub_Issue2654);
 	ZEPHIR_INIT(Stub_Issue2656);
 	ZEPHIR_INIT(Stub_Issue2666);
+	ZEPHIR_INIT(Stub_Issue2674);
 	ZEPHIR_INIT(Stub_Issue2679);
 	ZEPHIR_INIT(Stub_Issue663);
 	ZEPHIR_INIT(Stub_Issue760);